Move logic to annotate offense message. Out of Cop into its own class.
The Cop class was too large and failed style checks.
The moved logic pertains only to annotating the message
passed into an offense and not any other part of Cop.

Message annotation was not 100% tested, so attempted to move safely by:
1. Create loose scaffold spec in cop_spec.
2. Move logic into new class.
3. Move existing specs into new class's spec file.
4. Move scaffold specs into new class's spec file.
